+++ frontend-next/components/Game/constants.ts
@@ -1,1 +1,9 @@
 export const numberOfBombs = 20;
+
+export enum squareClassNames {
+  visible = "visible",
+  empty = "empty",
+  hidden = "hidden",
+  flag = "flag",
+  bomb = "bomb",
+}
